Add a JUnit Rule and base test case class for Spanner Tests
A JUnit test can extend the UsingSpannerEmulator class which will
provide a spanner emulated database with a given schema to the
testing class. The assumption is that the emulator is running
and the host is set via the $SPANNER_EMULATOR_HOST env variable.
This would be `export SPANNER_EMULATOR_HOST=localhost:9010` when
running on the same machine as the emulator with the default port
with

  gcloud beta emulators spanner start

The test class will create a new spanner instance in 'test-project'
and each @Test will create a new database in that insance so they
don't share state across tests. After each test the database is
dropped, and the @AfterClass drops the test-instance from the emulator.

An example of UsingSpannerEmulator is also given, although it is not
a comprehensive test of the database it does show two tests that insert
mutations into a databse using the computations.sdl schema file.

There is a shell script which can be the base of a shell_test that
will run the spanner emulator gcp docker image and remove it after
the test has finished. This way the developer of a test using the
spanner emulator can make a kt_jvm_binary() and a sh_test() rule
that depends on it and bazel test //:all will work and multiple spanner
emulator tests will be snadboxed to their own instances of the dokcer
containers. The spanner_emulator_test macro creates the kt_jvm_binary()
and sh_test() targets.

Change-Id: If5aecf715ebc7ee1820b7064ec4b72ef9813bed6