Implement /ExchangeStepAttempts.AppendLogEntry. (world-federation-of-advertisers/cross-media-measurement#18)
* Update cross-media-measurement-api version.

* Implement /ExchangeStepAttempts.AppendLogEntry.

This implements the v2alpha version.

See world-federation-of-advertisers/cross-media-measurement#3.

* Split apart KingdomRelationalDatabase. (world-federation-of-advertisers/cross-media-measurement#19)

This moves functionality into super-interfaces of
KingdomRelationalDatabase. This also splits out test-only logic into a
test-only database helper class.

The next steps are:

1. Split apart implementations of KingdomRelationalDatabase into
implementations of the super-interface.

2. Transition clients of KingdomRelationalDatabase to clients of the
super-interfaces.

3. Delete KingdomRelationalDatabase.

These changes are important because, as part of world-federation-of-advertisers/cross-media-measurement#3, a fair amount of
new database functionality is needed. However, most clients only need a
tiny subset of the exposed interface. By splitting it up, we are able to
logically isolate functionality and narrow the public interfaces seen by
most clients.

* Address comments to clean up ExchangeStepAttemptsServiceTest

* Fix typo in date-formatting string

* Spell out test constant rather than deriving it