Require DNS hostname to be specified for certificate generation macros.
This hostname gets set in the Subject Alternative Name (SAN) X.509 v3 extension of the generated certificate. Prior to this change, the value was hard-coded to "localhost".

This also moves the macros to the //build/openssl Bazel package to match the existing convention for build defs that call an external binary.