Pulling out grpc_extra_deps and updating the go_register_toolchains version (#189)
In the CMM repo, we want to add a GRPC Gateway for the BFF Reporting API to support the Reporting UI. GRPC Gateway requires to call go_register_toolchains(version = "1.19.4"). Or at least trying to change things to support the hardcoded version in the dependency would be difficult. The rest of the file is the same to minimize the risk of breaking anything.