build!: Upgrade Mockito to v5 (#263)
build!: Upgrade Mockito to v5

The capability to mock final methods is added in v5, which is necessary
for testing in some cases. This does, however, add the requirement to
target jvm version 11.

BREAKING CHANGE: Targets Java 11

To migrate to this commit, the .bazelrc file needs to include "build
--java_language_version=11". If a version of this already exists that
targets an older language version, update it to target 11. Otherwise,
add the line.