feat!: Allow a grace period for in-flight RPCs to complete (#310)
In particular, do not shut down the gRPC server executor explicitly prior to the grace period.

RELNOTES: gRPC servers now support a grace period to allow in-flight RPCs to complete when shutting down. This can be controlled by the `--shutdown-grace-period-seconds` option.
BREAKING CHANGE: `Instrumentation.instrumentThreadPool` returns a handle for cleanup instead of an `ExecutorService` that cleans up on shutdown.