fix: register aeadconfig before parsing (#320)
`val handle: KeysetHandle =
TinkProtoKeysetFormat.parseEncryptedKeyset(encryptedDek.toByteArray(),
kekAead, byteArrayOf())`

errors `Unsupported Key Type: LegacyProtoKey`.

That happens because the `StreamingAeadConfig` is registered after the
parser generates the `KeysetHandle`. So the parses fallback to
`LegacyProto` which is not supported.

This PR fixes the problem by registering the Config before the parser
runs.

---------

Co-authored-by: marcopremier <marcopremier@meta.com>