fix: Ensure that entire Spanner emulator STDERR stream is consumed (#351)
This appears to be what caused the problem that required reverting #346 in #350. The Spanner client library update required a new version of the Spanner emulator. This version included additional output to STDERR in some cases when PROTO BUNDLEs are used. The process would block on writing to STDERR without it being consumed.